Add template-grid class
I want to create a gap between rows of templates that is responsive. I planned
to make use of class .col and .row in layout.css but they only serve the space
between columns not rows so I had to create template.grid class so the rows can
be aligned well and the space gap stays responsive.